requirements.txt
discord.py[voice]
dotenv
lavalink
yt_dlp